switched home and "intro" (soon to be chatbot) pages
