+++ client-next/src/app/page.js
@@ -1,13 +1,20 @@
-import Link from 'next/link';
+import SlidingCardList from './components/SlidingCardList';
+import SlidingCardList2 from './components/SlidingCardList2';
+import './intro.css';
 
-export default function HomePage() {
+export default function IntroPage() {
   return (
-    <section>
-      <h1>Welcome to the Esports Manager</h1>
-      <p>Your AI-powered assistant for VALORANT esports team building and scouting.</p>
-      <Link href="/intro">
-        <button>Get Started</button>
-      </Link>
+    <section className="intro-page">
+      <div className="title">
+        <h1>VALORANT Esports Manager</h1>
+        <p>Build the best VALORANT team with AI-assisted analysis and scouting</p>
+      </div>
+      <SlidingCardList />
+      <SlidingCardList2 />
+      <SlidingCardList />
+      <div className="body">
+
+      </div>
     </section>
   );
 }
